feat: implement new challenge modal design
